Bug Fix : fixed the status bar and app bar issue in Android Lollipop 5.0.2
In brief, used fitsyStemWindows to positon the back arrow and menu (app
bar), added padding and internal flags to position it in between [width
center]  and then used translation in Y coordinate to correct the
position of the titile. Flags can only be set with the help of windows
translucent attribute.